Moved test and added parfor check
